feat: Show circle animation

+++ data/src/main/java/com/udacity/load/app/data/listener/ProgressListener.kt
@@ -2,6 +2,6 @@ package com.udacity.load.app.data.listener
 
 interface ProgressListener {
 
-    fun progress(value: Long)
+    fun progress(value: Int)
 
 }
